+++ src/OpenCqrs.EventSourcing.Store.Cosmos/CosmosDomainService.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Options;
-using OpenCqrs.EventSourcing.Configuration;
 using OpenCqrs.EventSourcing.Domain;
 using OpenCqrs.EventSourcing.Store.Cosmos.Configuration;
 using OpenCqrs.EventSourcing.Store.Cosmos.Documents;
@@ -26,11 +25,10 @@ public class CosmosDomainService : IDomainService
     /// Initializes a new instance of the <see cref="CosmosDomainService"/> class.
     /// </summary>
     /// <param name="cosmosOptions">Cosmos DB configuration options.</param>
-    /// <param name="eventSourcingOptions"></param>
     /// <param name="timeProvider">The time provider for timestamps.</param>
     /// <param name="httpContextAccessor">HTTP context accessor for user information.</param>
     /// <param name="cosmosDataStore">The Cosmos data store for document operations.</param>
-    public CosmosDomainService(IOptions<CosmosOptions> cosmosOptions, IOptions<EventSourcingOptions> eventSourcingOptions, TimeProvider timeProvider, IHttpContextAccessor httpContextAccessor, ICosmosDataStore cosmosDataStore)
+    public CosmosDomainService(IOptions<CosmosOptions> cosmosOptions, TimeProvider timeProvider, IHttpContextAccessor httpContextAccessor, ICosmosDataStore cosmosDataStore)
     {
         _timeProvider = timeProvider;
         _httpContextAccessor = httpContextAccessor;
@@ -58,12 +56,12 @@ public class CosmosDomainService : IDomainService
 
         if (aggregateDocumentResult.Value != null)
         {
-            var currentAggregate = aggregateDocumentResult.Value.ToAggregate<T>();
+            var currentAggregateDocument = aggregateDocumentResult.Value;
             if (!applyNewEvents)
             {
-                return currentAggregate;
+                return currentAggregateDocument.ToAggregate<T>();
             }
-            return await UpdateAggregate(streamId, aggregateId, cancellationToken);
+            return await _cosmosDataStore.UpdateAggregateDocument(streamId, aggregateId, currentAggregateDocument, cancellationToken);
         }
 
         var aggregate = new T();
@@ -610,11 +608,6 @@ public class CosmosDomainService : IDomainService
             return aggregateDocumentResult.Failure!;
         }
         var aggregateDocument = aggregateDocumentResult.Value;
-        if (aggregateDocument is null)
-        {
-            return new T();
-        }
-
         return await _cosmosDataStore.UpdateAggregateDocument(streamId, aggregateId, aggregateDocument, cancellationToken);
     }