+++ src/OpenCqrs.EventSourcing/Configuration/EventSourcingOptions.cs
@@ -1,8 +1,0 @@
-namespace OpenCqrs.EventSourcing.Configuration;
-
-public class EventSourcingOptions
-{
-    public bool StoreAggregateSnapshotByDefault { get; set; } = false;
-
-    public bool CheckForNewEventsByDefault { get; set; } = false;
-}