+++ docs/Release-Notes.md
@@ -2,7 +2,8 @@
 
 ## OpenCQRS 7.1.3
 _**Released 13/09/2025**_
-- Option to disable automatic snapshot creation in the Entity Framework Core and CosmosDB store providers
+- Fix issue when updating an aggregate that does not exist
+- Minor improvements
 
 ## OpenCQRS 7.1.2
 _**Released 10/09/2025**_